[Keymap]: rearrange modifier key
Using the tap key is very annoying. Sometimes it stuck.
Sometimes I press both keys too fast. So it interprets as a hold rather than a tap.
The timeout in config doesn't help much.

Other than that, I want to make all the modifiers consistent to use a sticky key.
So I don't need to remember which modifier is sticky and which one is not.
I try to press ctrl with my middle finger and see the result in the following days.